Refactored NexusFile and NexusNoFile
