+++ nexus-writer/src/nexus/mod.rs
@@ -1,16 +1,21 @@
+#[cfg(test)]
+mod mock_nexus_file;
+mod nexus_file;
+
 use std::path::Path;
+use hdf5::Group;
 
-use hdf5::{File, Group};
-
-use crate::{hdf5_handlers::NexusHDF5Result, nexus_structure::Root, run_engine::{
-    RunParameters,
+use crate::{hdf5_handlers::{ConvertResult, NexusHDF5Result}, run_engine::{
     run_messages::{
         InitialiseNewNexusStructure, PushAbortRunWarning, PushAlarm, PushFrameEventList,
         PushIncompleteFrameWarning, PushRunLogData, PushRunResumeWarning, PushRunStart,
         PushRunStop, PushSampleEnvironmentLog, SetEndTime,
-    },
+    }, GroupExt, RunParameters
 }, NexusSettings};
 
+#[cfg(test)]
+pub(crate) use mock_nexus_file::NexusNoFile;
+pub(crate) use nexus_file::NexusFile;
 
 pub(crate) trait NexusSchematic: Sized {
     const CLASS: &str;
@@ -84,64 +89,4 @@ pub(crate) trait NexusFileInterface:
     fn build_new_file(file_path: &Path, nexus_settings: &NexusSettings) -> NexusHDF5Result<Self>;
     fn open_from_file(file_path: &Path) -> NexusHDF5Result<Self>;
     fn extract_run_parameters(&self) -> NexusHDF5Result<RunParameters>;
-}
-
-pub(crate) struct NexusFile {
-    file: File,
-    root: Root,
-}
-
-impl NexusFileInterface for NexusFile {
-    fn build_new_file(file_path: &Path, nexus_settings: &NexusSettings) -> NexusHDF5Result<Self> {
-        let file = File::create(file_path)?;
-        let root = Root::build_group_structure(&file, nexus_settings)?;
-        Ok(Self { file, root })
-    }
-
-    fn open_from_file(file_path: &Path) -> NexusHDF5Result<Self> {
-        let file = File::create(file_path)?;
-        let root = Root::populate_group_structure(&file)?;
-        Ok(Self { file, root })
-    }
-
-    fn extract_run_parameters(&self) -> NexusHDF5Result<RunParameters> {
-        self.root.extract_run_parameters()
-    }
-}
-
-impl<M> NexusMessageHandler<M> for NexusFile
-where
-    Root: NexusMessageHandler<M>,
-{
-    fn handle_message(&mut self, message: &M) -> NexusHDF5Result<()> {
-        self.root.handle_message(message)
-    }
-}
-
-#[cfg(test)]
-pub(crate) struct NexusNoFile;
-
-#[cfg(test)]
-impl NexusFileInterface for NexusNoFile {
-    fn build_new_file(_: &Path, _: &NexusSettings) -> NexusHDF5Result<Self> {
-        Ok(Self)
-    }
-
-    fn open_from_file(_: &Path) -> NexusHDF5Result<Self> {
-        Ok(Self)
-    }
-
-    fn extract_run_parameters(&self) -> NexusHDF5Result<RunParameters> {
-        unreachable!()
-    }
-}
-
-#[cfg(test)]
-impl<M> NexusMessageHandler<M> for NexusNoFile
-where
-    Root: NexusMessageHandler<M>,
-{
-    fn handle_message(&mut self, _: &M) -> NexusHDF5Result<()> {
-        Ok(())
-    }
 }
